+++ Scripts/pi_s3cmd_put_enc.sh
@@ -5,6 +5,7 @@ export PATH=/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/bin
 SECRET_KEY=$1
 FILE_PATH=$2
 BUCKET=$3
+SCRIPT_PATH="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 # show help if needed
 if [[ $# -lt 3 ]]; then
@@ -13,7 +14,7 @@ if [[ $# -lt 3 ]]; then
 fi
 
 # transform the file (content and name) before uploading to S3
-ENC_FILE_PATH=$(pi_enc_file.sh $SECRET_KEY $FILE_PATH $PWD)
+ENC_FILE_PATH=$($SCRIPT_PATH/pi_enc_file.sh $SECRET_KEY $FILE_PATH $PWD)
 
 # upload to Amazon S3
 s3cmd put $ENC_FILE_PATH $BUCKET