+++ PICryptor/Sources/PICryptor.swift
@@ -8,6 +8,7 @@
 
 import Foundation
 import CommonCrypto
+import SwiftBase58
 
 fileprivate enum PICryptorError: Error {
     case unknownError
@@ -72,14 +73,14 @@ fileprivate enum PICryptorError: Error {
  * openssl encryption of a file content:
  * openssl rc4 -in <infile> -nosalt -K E86A53E1E6B5E1321615FD9FB90A7CAA  > <outfile>
  * 
- * encrypting a string and outputting a base64:
- * echo -n url_encoded_file_name | openssl rc4 -nosalt -K E86A53E1E6B5E1321615FD9FB90A7CAA | base64
+ * encrypting a string and outputting a base58:
+ * echo -n <infilename> | openssl rc4 -nosalt -K E86A53E1E6B5E1321615FD9FB90A7CAA | base58
  * 
  * decrypting contents of a file:
  * openssl rc4 -d -in <infile> -nosalt -K E86A53E1E6B5E1321615FD9FB90A7CAA  > <outfile>
  *
- * decyrpting a base64 filename:
- * echo -n url_decoded_file_name | base64 -D | openssl rc4 -d -nosalt -K E86A53E1E6B5E1321615FD9FB90A7CAA
+ * decyrpting a base58 filename:
+ * echo -n <infilename> | base58 -d | openssl rc4 -d -nosalt -K E86A53E1E6B5E1321615FD9FB90A7CAA
  */
 fileprivate enum RC4Cryptor {
     public static func decrypt(data: Data) -> Data? {
@@ -142,27 +143,24 @@ public extension Data {
 
 public extension NSString {
     /**
-     Get RC4+Base64 encrypted string representation.
+     Get RC4+Base58 encrypted string representation.
      */
-    public func rc4Base64Encrypted() -> NSString? {
-        if let data = self.data(using: String.Encoding.utf8.rawValue),
+    public func rc4Base58Encrypted() -> NSString? {
+        if let data = data(using: String.Encoding.utf8.rawValue),
             let encrypted = RC4Cryptor.encrypt(data: data) {
-            let base64 = encrypted.base64EncodedString()
-            return base64.RFC3986UnreservedEncoded as NSString?
+            let base58 = SwiftBase58.encode([UInt8](encrypted))
+            return base58 as NSString
         }
         return nil
     }
     
     /**
-     Get RC4+Base64 decrypted string representation.
+     Get RC4+Base58 decrypted string representation.
      */
-    public func rc4Base64Decrypted() -> NSString? {
-        if let string = self.removingPercentEncoding {
-            if let data = string.data(using: .utf8),
-                let base64 = Data(base64Encoded: data),
-                let decrypted = RC4Cryptor.decrypt(data: base64) {
-                return NSString(data: decrypted as Data, encoding: String.Encoding.utf8.rawValue)
-            }
+    public func rc4Base58Decrypted() -> NSString? {
+        let base58 = Data(bytes: SwiftBase58.decode(self as String))
+        if let decrypted = RC4Cryptor.decrypt(data: base58) {
+            return String(data: decrypted, encoding: .utf8)! as NSString
         }
         return nil
     }
@@ -170,34 +168,25 @@ public extension NSString {
 
 public extension String {
     
-    var RFC3986UnreservedEncoded: String? {
-        let unreservedChars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-._~"
-        let unreservedCharsSet: CharacterSet = CharacterSet(charactersIn: unreservedChars)
-        return addingPercentEncoding(withAllowedCharacters: unreservedCharsSet)
-    }
-    
     /**
-     Get RC4+Base64 encrypted string representation.
+     Get RC4+Base58 encrypted string representation.
      */
-    public func rc4Base64Encrypted() -> String? {
-        if let data = self.data(using: .utf8),
+    public func rc4Base58Encrypted() -> String? {
+        if let data = data(using: .utf8),
             let encrypted = RC4Cryptor.encrypt(data: data) {
-            let base64 = encrypted.base64EncodedString()
-            return base64.RFC3986UnreservedEncoded
+            let base58 = SwiftBase58.encode([UInt8](encrypted))
+            return base58
         }
         return nil
     }
     
     /**
-     Get RC4+Base64 decrypted string representation.
+     Get RC4+Base58 decrypted string representation.
      */
-    public func rc4Base64Decrypted() -> String? {
-        if let string = self.removingPercentEncoding {
-            if let data = string.data(using: .utf8),
-                let base64 = Data(base64Encoded: data),
-                let decrypted = RC4Cryptor.decrypt(data: base64) {
-                return String(data: decrypted, encoding: .utf8)
-            }
+    public func rc4Base58Decrypted() -> String? {
+        let base58 = Data(bytes: SwiftBase58.decode(self))
+        if let decrypted = RC4Cryptor.decrypt(data: base58) {
+            return String(data: decrypted, encoding: .utf8)
         }
         return nil
     }