+++ Pods/SkyS3Sync/SkyS3Sync/SkyS3SyncManager.m
@@ -101,10 +101,10 @@ static NSInteger const SkyS3MaxRetries = 3;
 #pragma mark - SkyResourceProvider
 
 - (NSURL *)URLForResource:(NSString *)name withExtension:(NSString *)ext {
-    if (!name || !ext) {
+    if (!name) {
         return nil;
     }
-    NSString *resourceFileName = [name stringByAppendingPathExtension:ext];
+    NSString *resourceFileName = ext.length > 0 ? [name stringByAppendingPathExtension:ext] : name;
     return [self URLForResourceWithFileName:resourceFileName];
 }
 
@@ -288,10 +288,12 @@ static NSInteger const SkyS3MaxRetries = 3;
 
     @weakify(self);
     void(^completedBlock)() = ^{
-        if (++completedCounter == remoteResources.count) {
-            @strongify(self);
-            [self finishSync];
-        };
+        @strongify(self);
+        @synchronized(self) {
+            if (++completedCounter == remoteResources.count) {
+                [self finishSync];
+            };
+        }
     };
 
     // delete local files that are absent on the remote host