+++ Scripts/pi_enc_filename.sh
@@ -12,5 +12,4 @@ if [[ $# -lt 2 ]]; then
 fi
 
 # return encrypted file name
-DATA=`echo -n $FILE_NAME | openssl rc4 -nosalt -K $SECRET_KEY | base64`
-echo $(python -c "import sys, urllib as ul; print ul.quote_plus('$DATA')")
+echo `echo -n $FILE_NAME | openssl rc4 -nosalt -K $SECRET_KEY | base58`