+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 CHANGELOG
 =========
 
+## pi_1.5
+determine the path of the executing BASH script
+
 ## pi_1.4
 scripts should accept key as a first param