+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 CHANGELOG
 =========
 
-## 1.5
+## pi_1.6
+use url encode/decode for file names
+
+## pi_1.5
 determine the path of the executing BASH script
 
 ## 1.4